ContentManagerImp moved to .clear sub-package.
