+++ dtglib/src/main/java/com/kaltura/dtg/ContentManager.java
@@ -2,6 +2,8 @@ package com.kaltura.dtg;
 
 import android.content.Context;
 
+import com.kaltura.dtg.clear.ContentManagerImp;
+
 import java.io.File;
 import java.util.List;
